btoa handling a string with nulls or with nonascii characters

+++ src/ebprot.h
@@ -79,7 +79,7 @@ bool looksBinary(const unsigned char *buf, int buflen);
 void looks_8859_utf8(const uchar *buf, int buflen, bool * iso_p, bool * utf8_p);
 void iso2utf(const uchar *inbuf, int inbuflen, uchar **outbuf_p, int *outbuflen_p);
 void utf2iso(const uchar *inbuf, int inbuflen, uchar **outbuf_p, int *outbuflen_p);
-void utf2iso1(char *s, int *lenp);
+void utf2iso1(char *s, size_t *lenp);
 void utfHigh(const char *inbuf, int inbuflen, char **outbuf_p, int *outbuflen_p, bool inutf8, bool out32, bool outbig);
 char *uni2utf8(unsigned int unichar);
 void utfLow(const char *inbuf, int inbuflen, char **outbuf_p, int *outbuflen_p, int bom);
